be/requirements.txt
flask
flask-cors
python-dotenv
supabase
flask-jwt-extended
pydantic-settings